+++ src/main/webapp/WEB-INF/config/adminConfig/adminCSBoard-servlet.xml
@@ -22,5 +22,5 @@
 		<property name="suffix" value=".jsp" />
 	</bean>
 
-
+	<context:component-scan base-package="com.team5.funthing"></context:component-scan>
 </beans>
