+++ src/main/webapp/WEB-INF/web.xml
@@ -88,8 +88,8 @@
 		<init-param>
 			<param-name>contextConfigLocation</param-name>
 			<param-value>/WEB-INF/config/adminConfig/admin-servlet.xml</param-value>
-
 			<!-- 추가할 admin 설정 파일 -->
+			<param-value>/WEB-INF/config/adminConfig/adminCSBoard-servlet.xml</param-value>
 		</init-param>
 		<load-on-startup>1</load-on-startup>
 	</servlet>
