+++ src/main/webapp/resources/common/ckeditor5/js/ckeditor.js
